requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-openai
langchain-community
sqlalchemy
psycopg2-binary
chromadb
tiktoken
